# octbase-api/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Octbase API
  version: "0.4.0"
  description: |
    REST API for Octbase — a focused project-management tool (tasks, boards,
    releases, sprints, wiki pages, SCM integration) for a single, invitation-only
    team.

    **Authentication:** All routes require a JWT **Bearer** access token in the
    `Authorization` header, except the public auth/invitation routes, the
    HMAC-authenticated webhook receivers, and the unauthenticated ops endpoints
    (`/api/v1/health`, `/api/v1/version`, `/api/v1/meta/enums`, `/api/v1/config`, `/metrics`,
    `/openapi.yaml`, `/docs`). Obtain an access token from `POST /api/v1/auth/login`;
    it is returned as `{ "accessToken", "expiresIn" }`. A long-lived refresh token is
    set as an HttpOnly cookie and rotated on every `POST /api/v1/auth/refresh`.

    **Authorization:** Two layers, both enforced server-side. A global role
    (`SUPER_ADMIN`, `ADMIN`, `USER`, `GUEST`) and a per-project membership role
    (`PROJECT_OWNER`, `PROJECT_ADMIN`, `PROJECT_MEMBER`, `PROJECT_VIEWER`). The
    effective project permission matrix is available at
    `GET /api/v1/projects/{projectId}/permissions`.

    **Error format:** All errors return
    `{ "code": "ERROR_CODE", "message": "...", "messageKey": "...", "details": {...} }`
    (`messageKey`/`details` optional).

    **Pagination:** List endpoints accept `?page=0&size=20` (zero-indexed pages,
    default size 20).

...
  contact:
    name: Octbase
...
servers:
  - url: /
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    projectId: { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        messageKey: { type: string }
        details: { type: object, additionalProperties: true }
    CountDistribution:
      type: array
      description: >
        One bucket per enum value, in the domain's own presentation order and
        including the empty ones, so a chart drawn from it keeps its shape
        between reloads. Values outside the built-in set (a project's custom
        priority, a board-lane status) follow the known ones alphabetically
        rather than being dropped.
      items:
        type: object
        properties:
          key: { type: string }
          count: { type: integer }
...
paths:
...
  /api/v1/projects/{projectId}/reports/statistics:
    parameters: [ { $ref: '#/components/parameters/projectId' } ]
    get:
      tags: [Reports]
      summary: Project statistics overview for a project manager
      description: >
        Everything the statistics page draws, in one read-time projection —
        no snapshot table and no scheduled job. Task totals and the
        status/type/priority distributions (empty buckets included, in the
        domain's own order, so a chart keeps its shape between reloads);
        overdue, due-soon and unassigned counts over OPEN tasks only; the
        30-day created/completed flow; weekly throughput for the last 8
        weeks; average and median cycle time from creation to DONE over the
        last 90 days; open work per assignee; the active sprint; and the
        release plan. `effort` is null unless the project estimates, so zeros
        can never be mistaken for "nothing left to do". Requires project
        membership — the numbers aggregate exactly what a member can already
        read task by task.
      operationId: projectStatistics
      responses:
        '200':
          description: Project statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId: { type: string, format: uuid }
                  generatedAt: { type: string, format: date-time }
                  estimationUnit: { type: string, enum: [NONE, POINTS, HOURS] }
                  tasks:
                    type: object
                    properties:
                      total: { type: integer }
                      open: { type: integer }
                      inProgress: { type: integer }
                      done: { type: integer }
                      archived: { type: integer }
                      unassigned: { type: integer }
                      overdue: { type: integer }
                      dueSoon: { type: integer }
                      createdLast30: { type: integer }
                      completedLast30: { type: integer }
                      byStatus: { $ref: '#/components/schemas/CountDistribution' }
                      byType: { $ref: '#/components/schemas/CountDistribution' }
                      byPriority: { $ref: '#/components/schemas/CountDistribution' }
                  effort:
                    type: object
                    nullable: true
                    properties:
                      unit: { type: string, enum: [POINTS, HOURS] }
                      total: { type: number }
                      done: { type: number }
                      remaining: { type: number }
                      unestimated: { type: integer }
                  throughput:
                    type: array
                    items:
                      type: object
                      properties:
                        weekStart: { type: string, format: date }
                        completed: { type: integer }
                        effort: { type: number, nullable: true }
                  cycleTime:
                    type: object
                    properties:
                      sampleSize: { type: integer }
                      averageDays: { type: number, nullable: true }
                      medianDays: { type: number, nullable: true }
                  workload:
                    type: array
                    items:
                      type: object
                      properties:
                        userId: { type: string, format: uuid }
                        name: { type: string }
                        open: { type: integer }
                        effort: { type: number, nullable: true }
                  sprint:
                    type: object
                    nullable: true
                    properties:
                      sprintId: { type: string, format: uuid }
                      name: { type: string }
                      startDate: { type: string, format: date, nullable: true }
                      endDate: { type: string, format: date, nullable: true }
                      committed: { type: integer }
                      completed: { type: integer }
                      daysRemaining: { type: integer, nullable: true, description: End date inclusive — a sprint ending today has 1 day left }
                      committedEstimate: { type: number, nullable: true }
                      completedEstimate: { type: number, nullable: true }
                  releases:
                    type: object
                    properties:
                      open: { type: integer }
                      closed: { type: integer }
                      nextDue: { type: string, format: date, nullable: true }
                      nextDueName: { type: string, nullable: true }
                      overdueOpen: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────── Task categories & templates ───────────────
...